+++ packages/jaspr_riverpod/lib/experimental/persist.dart
@@ -1,0 +1,8 @@
+export '../src/internals.dart'
+    show
+        StorageOptions,
+        PersistedData,
+        Storage,
+        StorageCacheTime,
+        NotifierPersistX,
+        PersistResult;