+++ packages/jaspr_riverpod/lib/src/core/provider_scope.dart
@@ -1,21 +1,11 @@
 // ignore_for_file: invalid_use_of_internal_member
-
-import 'dart:async';
-
-import 'package:jaspr/jaspr.dart';
-import 'package:meta/meta.dart';
-
-import 'internals.dart';
-import 'sync_provider.dart';
-
-part 'provider_context.dart';
-part 'provider_dependencies.dart';
+part of '../core.dart';
 
 /// {@template riverpod.provider_scope}
 /// A component that stores the state of providers.
 ///
 /// All Jaspr applications using Riverpod must contain a [ProviderScope] at
-/// the root of their component tree. It is done as followed:
+/// the root of their client-side component tree. It is done as followed:
 ///
 /// ```dart
 /// void main() {
@@ -28,12 +18,12 @@ part 'provider_dependencies.dart';
 /// }
 /// ```
 ///
-/// It optionally possible to specify `overrides` to change the behavior of
+/// It's optionally possible to specify `overrides` to change the behavior of
 /// some providers. This can be useful for testing purposes:
 ///
 /// ```dart
 /// testComponents('Test example', (tester) async {
-///   tester.pumpComponent(
+///   await tester.pumpComponent(
 ///     ProviderScope(
 ///       overrides: [
 ///         // override the behavior of repositoryProvider to provide a fake
@@ -48,7 +38,7 @@ part 'provider_dependencies.dart';
 ///
 ///
 /// Similarly, it is possible to insert other [ProviderScope] anywhere inside
-/// the component tree to override the behavior of a provider for only a part of the
+/// the widget tree to override the behavior of a provider for only a part of the
 /// application:
 ///
 /// ```dart
@@ -57,17 +47,18 @@ part 'provider_dependencies.dart';
 /// void main() {
 ///   runApp(
 ///     ProviderScope(
-///       child: MaterialApp(
-///         // Home uses the default behavior for all providers.
-///         home: Home(),
-///         routes: {
+///       child: Router(
+///         routes: [
 ///           // Overrides themeProvider for the /gallery route only
-///           '/gallery': (_) => ProviderScope(
-///             overrides: [
-///               themeProvider.overrideWithValue(MyTheme.dark()),
-///             ],
+///           Route(
+///             path: '/gallery',
+///             builder: (_) => ProviderScope(
+///               overrides: [
+///                 themeProvider.overrideWithValue(MyTheme.dark()),
+///               ],
+///             ),
 ///           ),
-///         },
+///         ],
 ///       ),
 ///     ),
 ///   );
@@ -76,17 +67,18 @@ part 'provider_dependencies.dart';
 ///
 /// See also:
 /// - [ProviderContainer], a Dart-only class that allows manipulating providers
-/// - [UncontrolledProviderScope], which exposes a [ProviderContainer] to the component
+/// - [UncontrolledProviderScope], which exposes a [ProviderContainer] to the widget
 ///   tree without managing its life-cycles.
 /// {@endtemplate}
-@sealed
-class ProviderScope extends StatefulComponent {
+/// {@category Core}
+final class ProviderScope extends StatefulComponent {
   /// {@macro riverpod.provider_scope}
   const ProviderScope({
     super.key,
+    this.sync = const [],
     this.overrides = const [],
     this.observers,
-    this.parent,
+    this.retry,
     required this.child,
   });
 
@@ -110,117 +102,130 @@ class ProviderScope extends StatefulComponent {
     return element;
   }
 
-  /// Explicitly override the parent [ProviderContainer] that this [ProviderScope]
-  /// would be a descendant of.
+  /// The retry logic used by providers associated to this container.
   ///
-  /// A common use-case is to allow modals to access scoped providers, as they
-  /// would otherwise be unable to since they would be in a different branch
-  /// of the component tree.
-  ///
-  /// That can be achieved with:
-  ///
-  /// ```dart
-  /// ElevatedButton(
-  ///   onTap: () {
-  ///     final container = ProviderScope.containerOf(context);
-  ///     showDialog(
-  ///       context: context,
-  ///       builder: (context) {
-  ///         return ProviderScope(parent: container, child: MyModal());
-  ///       },
-  ///     );
-  ///   },
-  ///   child: Text('show modal'),
-  /// )
-  /// ```
-  ///
-  ///
-  /// The [parent] variable must never change.
-  final ProviderContainer? parent;
+  /// See [ProviderContainer.defaultRetry] for information about the
+  /// default retry logic.
+  final Retry? retry;
 
-  /// The part of the component tree that can use Riverpod and has overridden providers.
+  /// The part of the widget tree that can use Riverpod and has overridden providers.
   final Component child;
 
   /// The listeners that subscribes to changes on providers stored on this [ProviderScope].
   final List<ProviderObserver>? observers;
 
   /// Information on how to override a provider/family.
+  ///
+  /// Overrides are created using methods such as [Provider.overrideWith]/[Provider.overrideWithValue].
+  ///
+  /// This can be used for:
+  /// - testing, by mocking a provider.
+  /// - dependency injection, to avoid having to pass a value to many
+  ///   widgets in the widget tree.
+  /// - performance optimization: By using this to inject values to widgets
+  ///   using `ref` inside of their constructor, widgets may be able to use
+  ///   `const` constructors, which can improve performance.
+  ///
+  /// **Note**: Overrides only apply to this [ProviderScope] and its descendants.
+  /// Ancestors of this [ProviderScope] will not be affected by the overrides.
   final List<Override> overrides;
+
+  final List<ProviderSync> sync;
 
   @override
   ProviderScopeState createState() => ProviderScopeState();
 }
 
 /// Do not use: The [State] of [ProviderScope]
-@visibleForTesting
-@sealed
-class ProviderScopeState extends State<ProviderScope> with SyncScopeMixin {
+@internal
+final class ProviderScopeState extends State<ProviderScope> with SyncScopeMixin {
   /// The [ProviderContainer] exposed to [ProviderScope.child].
   @override
   @visibleForTesting
-  // ignore: diagnostic_describe_all_properties
   late final ProviderContainer container;
-
   ProviderContainer? _debugParentOwner;
   var _dirty = false;
 
+  bool didInitContainer = false;
+
+  @override
+  Future<void> preloadState() {
+    _initContainer();
+    return _preloadSyncProviders();
+  }
+
   @override
   void initState() {
+    super.initState();
+    assert(!kIsWeb || !didInitContainer, 'Container should not be pre-initialized on the web.');
+    _initContainer();
+  }
+
+  void _initContainer() {
+    if (didInitContainer) return;
+    didInitContainer = true;
+
     final parent = _getParent();
-    assert(() {
+    if (kDebugMode) {
       _debugParentOwner = parent;
+    }
+
+    assert(() {
+      if (parent == null) return true;
+
+      for (final s in component.sync) {
+        if ((s as _ProviderSync).provider.dependencies == null) {
+          print('[WARNING] Tried to sync provider "${s.provider.name ?? s.provider.runtimeType}" on a nested ProviderScope, but the provider did not specify dependencies.\n'
+            '  Because syncing a provider will override it on the client, and overriding in a nested ProviderScope requires the provider to specify dependencies, the sync will fail.\n'
+            '  To fix this, either specify dependencies for the provider (e.g. `Provider(..., dependencies: [])`) and its dependents, or sync it on the root ProviderScope.',
+          );
+        }
+      }
+
       return true;
-    }(), '');
+    }());
 
     container = ProviderContainer(
       parent: parent,
       overrides: [
         _bindingProvider.overrideWithValue(context.binding),
+        ..._syncOverrides,
         ...component.overrides,
       ],
       observers: component.observers,
+      retry: component.retry,
+      // TODO: handle errors
+      // onError: (err, stack) {},
     );
-
-    super.initState();
   }
 
   ProviderContainer? _getParent() {
-    if (component.parent != null) {
-      return component.parent;
-    } else {
-      final scope = context.getElementForInheritedComponentOfExactType<UncontrolledProviderScope>()?.component
-          as UncontrolledProviderScope?;
+    final scope = context.getElementForInheritedComponentOfExactType<UncontrolledProviderScope>()?.component
+        as UncontrolledProviderScope?;
 
-      return scope?.container;
-    }
+    return scope?.container;
   }
 
   @override
   void didUpdateComponent(ProviderScope oldComponent) {
     super.didUpdateComponent(oldComponent);
     _dirty = true;
+  }
 
-    if (oldComponent.parent != component.parent) {
-      print("ERROR: Changing ProviderScope.parent is not supported");
+  void _debugAssertParentDidNotChange() {
+    final parent = _getParent();
+
+    if (parent != _debugParentOwner) {
+      throw UnsupportedError(
+        'ProviderScope was rebuilt with a different ProviderScope ancestor',
+      );
     }
   }
 
   @override
   Component build(BuildContext context) {
-    assert(() {
-      if (component.parent != null) {
-        // didUpdateComponent already takes care of component.parent change
-        return true;
-      }
-      final parent = _getParent();
+    if (kDebugMode) _debugAssertParentDidNotChange();
 
-      if (parent != _debugParentOwner) {
-        throw UnsupportedError(
-          'ProviderScope was rebuilt with a different ProviderScope ancestor',
-        );
-      }
-      return true;
-    }(), '');
     if (_dirty) {
       _dirty = false;
       container.updateOverrides([
@@ -229,10 +234,7 @@ class ProviderScopeState extends State<ProviderScope> with SyncScopeMixin {
       ]);
     }
 
-    return UncontrolledProviderScope(
-      container: container,
-      child: component.child,
-    );
+    return UncontrolledProviderScope(container: container, child: component.child);
   }
 
   @override
@@ -245,9 +247,13 @@ class ProviderScopeState extends State<ProviderScope> with SyncScopeMixin {
 /// {@template riverpod.UncontrolledProviderScope}
 /// Expose a [ProviderContainer] to the component tree.
 ///
-/// This is what makes `ref.watch(`/`Consumer`/`ref.read` work.
+/// **Note**: The [container] will _not_ be disposed when using this component.
+/// It is the caller's responsibility to dispose of it when no longer needed.
+/// Alternatively, use [ProviderScope] to automatically manage the lifecycle of
+/// the [ProviderContainer].
+///
 /// {@endtemplate}
-@sealed
+/// {@category Core}
 class UncontrolledProviderScope extends InheritedComponent {
   /// {@macro riverpod.UncontrolledProviderScope}
   const UncontrolledProviderScope({
@@ -265,8 +271,7 @@ class UncontrolledProviderScope extends InheritedComponent {
   }
 
   @override
-  // ignore: library_private_types_in_public_api
-  _UncontrolledProviderScopeElement createElement() {
+  InheritedElement createElement() {
     return _UncontrolledProviderScopeElement(this);
   }
 }
@@ -296,7 +301,7 @@ class _UncontrolledProviderScopeElement extends InheritedElement {
     getDependencies(dependent)!.listen(target, listener, onError: onError, fireImmediately: fireImmediately);
   }
 
-  ProviderSubscription<T> _subscribe<T>(
+  ProviderSubscription<T> _listenManual<T>(
     ProviderListenable<T> provider,
     void Function(T? previous, T value) listener, {
     void Function(Object error, StackTrace stackTrace)? onError,