+++ packages/jaspr_riverpod/CHANGELOG.md
@@ -1,3 +1,26 @@
+## Unreleased minor
+
+- Upgrade `riverpod` to `^3.0.0`
+
+- **Breaking**: Removed `SyncProvider`. Migrate to `ProviderScope(sync: [...])`.
+
+  Added `sync` option to `ProviderScope`, which allows to specify providers that should sync its value to the client using:
+
+  ```dart
+  ProviderScope(
+    sync: [
+      myProvider.syncWith('my-unique-id'),
+    ],
+    child: ...
+  )
+  ```
+
+  The `.syncWith` extension is available on `Provider`, `FutureProvider`, `StateProvider`, `NotifierProvider` and `AsyncNotifierProvider`.
+
+  For `FutureProvider` and `AsyncNotifierProvider` this also awaits the future during pre-rendering before building the child component.
+
+- **Breaking**: Renamed `context.subscribe()` to `context.listenManual()`.
+
 ## 0.3.22-dev.1
 
 - `jaspr` upgraded to `0.21.0-dev.1`